fix int overflow

+++ src/parcsr_ls/par_krylov_func.c
@@ -72,7 +72,7 @@ hypre_ParKrylovCreateVectorArray(HYPRE_Int n, void *vvector )
 
    size = hypre_VectorSize(hypre_ParVectorLocalVector(vector));
    num_vectors = hypre_VectorNumVectors(hypre_ParVectorLocalVector(vector));
-   array_data = hypre_CTAlloc(HYPRE_Complex, (n * size * num_vectors), memory_location);
+   array_data = hypre_CTAlloc(HYPRE_Complex, ((size_t) n) * size * num_vectors, memory_location);
    new_vector = hypre_CTAlloc(hypre_ParVector*, n, HYPRE_MEMORY_HOST);
    for (i = 0; i < n; i++)
    {
